fixes to resume/continue-pact - makes it impossible to test without a current calling module
