+++ tests/pact/yield.repl
@@ -71,22 +71,3 @@
 (env-entity "B")
 (continue-pact 0 true)
 (expect "step 0 rollback skips B" false (at "executed" (pact-state)))
-
-(commit-tx)
-(begin-tx)
-
-(use yield-test)
-
-(env-hash (hash "stu-pact-id"))
-
-;; execute 'tester-2' on chain 0
-(env-chain-data { "chain-id": "0" })
-(env-entity "A")
-(expect "step 0 executes" true (and (= (tester-2 "emily") "emily->A") (= (at 'chain-id (chain-data)) "0")))
-
-;; expect failure when yielding to the wrong chain id
-;; set resume for step 1, expect success in correct chain env
-(env-entity "B")
-(expect-failure "step 1 fails to execute on chain 0" (continue-pact 1))
-(env-chain-data { "chain-id": "1" })
-(expect "step 1 execution succeeds on correct chain id" "emily->A->B" (continue-pact 1))
